requirements.txt
boto3~=1.28.5
cryptography~=41.0.2
aws-lambda-powertools
PyJWT~=2.8.0
botocore~=1.31.5
jmespath~=1.0.1
s3transfer~=0.6.1
six~=1.16.0
python-dateutil~=2.8.2
urllib3~=1.26.16